requirements.txt
fastapi
pydantic
openai
python-dotenv
slowapi
pydantic-settings
uvicorn
typing
httpx
